requirements.txt
gradio==5.0.1
nltk==3.8.1
Pillow==9.4.0
Pillow==10.4.0
wordcloud==1.9.3
